Updated the attendance functions, and calls

+++ drizzle/schema.ts
@@ -1,5 +1,6 @@
 import { mysqlTable, mysqlSchema, AnyMySqlColumn, varchar, int, date, timestamp, datetime, text, mediumtext } from "drizzle-orm/mysql-core"
 import { sql } from "drizzle-orm"
+import { boolean } from "drizzle-orm/gel-core";
 
 export const attendanceRecord = mysqlTable("attendance_record", {
 	eNumber: varchar({ length: 10 }).notNull(),
@@ -14,7 +15,7 @@ export const classes = mysqlTable("classes", {
 	teacher: int().default('NULL'),
 	name: varchar({ length: 50 }).notNull(),
 	description: varchar({ length: 500 }).notNull(),
-	recurring: tinyint().notNull(),
+	recurring: int().notNull(),
 	rosterFile: varchar("roster_file", { length: 1000 }).default('NULL'),
 });
 
@@ -24,14 +25,14 @@ export const meetings = mysqlTable("meetings", {
 	location: varchar({ length: 30 }).notNull(),
 	date: datetime({ mode: 'string'}).default('NULL'),
 	qrcode: text().default('NULL'),
-	cancelled: tinyint().default(0),
+	cancelled: int().default(0),
 });
 
 export const users = mysqlTable("users", {
 	id: int().autoincrement().notNull(),
 	eNumber: varchar({ length: 10 }).notNull(),
 	passwordHash: mediumtext("password_hash").notNull(),
-	admin: tinyint().notNull(),
+	admin: int().notNull(),
 	firstName: varchar("first_name", { length: 20 }).notNull(),
 	lastName: varchar("last_name", { length: 20 }).notNull(),
 });
